third_party/blink/web_tests/platform/linux/virtual/webnn-service-with-gpu/external/wpt/webnn/conformance_tests/dequantizeLinear.https.any_gpu-expected.txt

This is a testharness.js-based test.
[FAIL] [required] dequantizeLinear int8 0D tensor with float32 0D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual 134.5 should be close enough to expected 134.43296813964844 by ULP distance: expected a number less than or equal to 1n but got 4393n
[FAIL] [required] dequantizeLinear int8 0D constant tensor with float32 0D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual 134.375 should be close enough to expected 134.43296813964844 by ULP distance: expected a number less than or equal to 1n but got 3799n
[FAIL] [required] dequantizeLinear uint8 1D constant tensor with float32 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -1084 should be close enough to expected -1083.798828125 by ULP distance: expected a number less than or equal to 1n but got 1648n
[FAIL] [required] dequantizeLinear uint8 1D constant tensor with negative float32 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -1084 should be close enough to expected -1083.798828125 by ULP distance: expected a number less than or equal to 1n but got 1648n
[FAIL] [required] dequantizeLinear uint8 1D constant tensor with float32 1D scale, implicit block_size = 2
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -1084 should be close enough to expected -1083.798828125 by ULP distance: expected a number less than or equal to 1n but got 1648n
[FAIL] [required] dequantizeLinear int8 4D constant tensor broadcasting float32 4D scale and int8 4D zeroPoint
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -38.09375 should be close enough to expected -38.08934783935547 by ULP distance: expected a number less than or equal to 1n but got 1154n
[FAIL] [required] dequantizeLinear int8 2D constant tensor with float32 2D scale, block_size = [3, 2]
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -35 should be close enough to expected -35.00859069824219 by ULP distance: expected a number less than or equal to 1n but got 2252n
[FAIL] [optional] dequantizeLinear int4 1D tensor of even size with float32 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -8.9609375 should be close enough to expected -8.962198257446289 by ULP distance: expected a number less than or equal to 1n but got 1322n
[FAIL] [optional] dequantizeLinear int4 1D tensor of odd size with float32 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual 2.240234375 should be close enough to expected 2.2405495643615723 by ULP distance: expected a number less than or equal to 1n but got 1322n
[FAIL] [optional] per-tensor dequantizeLinear for int4 4D constant with float32 4D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -2.240234375 should be close enough to expected -2.2405495643615723 by ULP distance: expected a number less than or equal to 1n but got 1322n
[FAIL] [required] dequantizeLinear int32 1D tensor with float32 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float32 actual -22400 should be close enough to expected -22405.495643615723 by ULP distance: expected a number less than or equal to 1n but got 2814n
[FAIL] [required] quantizeLinear then dequantizeLinear with different float32 scale and int8 zeroPoint
  assert_less_than_equal: assert_array_approx_equals_ulp: test quantizeLinear dequantizeLinear float32 actual -35.84375 should be close enough to expected -35.848793029785156 by ULP distance: expected a number less than or equal to 1n but got 1322n
[FAIL] [optional] dequantizeLinear uint8 1D constant tensor with float16 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float16 actual -972 should be close enough to expected -29.125 by ULP distance: expected a number less than or equal to 1 but got 5200
[FAIL] [optional] dequantizeLinear uint8 1D constant tensor with negative float16 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float16 actual -972 should be close enough to expected -29.125 by ULP distance: expected a number less than or equal to 1 but got 5200
[FAIL] [optional] dequantizeLinear int32 1D tensor with float16 1D scale
  assert_less_than_equal: assert_array_approx_equals_ulp: test dequantizeLinear float16 actual Infinity should be close enough to expected 18336 by ULP distance: expected a number less than or equal to 1 but got 1926
[FAIL] [optional] quantizeLinear then dequantizeLinear with different float16 scale and int8 zeroPoint
  assert_less_than_equal: assert_array_approx_equals_ulp: test quantizeLinear dequantizeLinear float16 actual 14.5625 should be close enough to expected 3.640625 by ULP distance: expected a number less than or equal to 1 but got 2048
Harness: the test ran to completion.
